rename covenant-identity file
